+++ cloudamqp/resource_cloudamqp_integration_metric.go
@@ -272,8 +272,9 @@ func (r *integrationMetricResource) Read(ctx context.Context, req resource.ReadR
 		return
 	}
 
-	// Handle resource drift and trigger re-creation if resource been deleted outside the provider
+	// Resource drift: instance or resource not found, trigger re-creation
 	if data == nil {
+		tflog.Info(ctx, fmt.Sprintf("metric integration not found, resource will be recreated: %s", state.ID.ValueString()))
 		resp.State.RemoveResource(ctx)
 		return
 	}